requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
deprecation==2.1.0
Flask==2.3.3
idna==3.4
packaging==23.1
pdpyras==5.1.2
python-dotenv==1.0.0
pytest==7.4.2
requests==2.31.0
urllib3==2.0.5
Flask-Cors==4.0.0